fix this-read-syntax use of quasisyntax/loc
The src argument to read-syntax is not a syntax object, so it is
not a suitable loc argument to {quasi,}syntax/loc. This used to work,
because syntax/loc ignored the argument if the template was just a
pattern variable (or for quasisyntax/loc, an unquote form). But the
new syntax form does stricter checking.