support external queries in `#lang datalog`
Support for external queries makes `#lang datalog` better for demos.

Supporting externals in the non-S-expression syntax requires adding a
way to import bindings and adding a way to call externals (with the
caveat that externals can break termination). Externals work better if
the obvious subset of literals are interpreted as integers and
booleans.

This patch makes those additions without adding new keywords, and it
should not break any existing `#lang datalog` programs. It could
change the interpretation of existing `#lang datalog` programs to use
numbers and booleans in place of symbols, but the syntax of numbers is
constrained so that formerly distinct symbols will not be collapsed to
the same number or boolean.